// Task_1/src/models/Transaction.ts
export interface Transaction {
   stt: number;
   ngay: string;
   gio: string;
   tram: string;
   tru_bom: string;
   mat_hang: string;
   so_luong: number;
   don_gia: number;
   thanh_tien: number;
   trang_thai_TT: string;
   ma_KH: string;
   ten_KH: string;
   loai_KH: string;
   ngay_thanh_toan: string;
   nhan_vien: string;
   bien_so_xe: string;
   trang_thai_HD: string;
}